+++ src/networking/message_processor.h
@@ -30,7 +30,7 @@ class MessageProcessor : public Singleton<MessageProcessor> {
         void handleNewLightColors(NewLightColors newLightColors);
         void handleObstacleAvoidanceCommand(ObstacleAvoidanceStatus status);
 
-        void processBinaryMessage(const uint8_t* data, size_t length);
+        void processBinaryMessage(const uint8_t* data, uint16_t length);
     private:
         void updateMotorSpeeds(int16_t leftSpeed, int16_t rightSpeed);
         void executeCommand(int16_t leftSpeed, int16_t rightSpeed);
